Conflicts:
	src/imports/controls/Menu.qml
	src/imports/controls/fusion/Menu.qml
	src/imports/controls/imagine/Menu.qml
	src/imports/controls/material/Menu.qml
	src/imports/controls/universal/Menu.qml

+++ src/quicktemplates2/qquicktumbler.cpp
@@ -867,7 +867,7 @@ void QQuickTumblerAttachedPrivate::init(QQuickItem *delegateItem)
 
 void QQuickTumblerAttachedPrivate::calculateDisplacement()
 {
-    const int previousDisplacement = displacement;
+    const qreal previousDisplacement = displacement;
     displacement = 0;
 
     if (!tumbler) {
